Make PFDA vignette not stop on error
Running on CentOS 7 with R 4.1.2 with fully-updated packages on 2022 March 17
produced the following error
```
   --- re-building ‘faustPFDA.Rmd’ using rmarkdown
   Quitting from lines 426-439 (faustPFDA.Rmd) 
   Error: processing vignette 'faustPFDA.Rmd' failed with diagnostics:
   Downdated VtV is not positive definite
   --- failed re-building ‘faustPFDA.Rmd’
``` 

when building vignettes on install, which blocked
the install. To allow the Rmd to complete even with the error (and allow at least to see part of that vignette and the other two vignettes
when installing with vignettes), I added
`error = TRUE` as an argument to `knitr::opts_chunk$set`.