Fix modVals logical comparison length bug
In R3.6.3, using vectors with length greater than one in the condition of an if statement merely warned and used the first element, but past verion 4.something it throws an error. 

It's not entirely clear to me whether any or all would be better. Presumably any, as if there are any NAs then we would want to fix the behaviour. So that's what we went with.